fix: グラフの最大値（Y軸）を総ページ数に固定し、データ生成処理をConcernに共通化 (#290)
