+++ .steering/20260605-issue290-chart-max-fixed/requirements.md
@@ -1,0 +1,34 @@
+# 要求内容
+
+## 概要
+
+詳細画面の読書進捗グラフのY軸（最大値）を、現在の読書進捗に関わらず、最初から対象書籍の「総ページ数」に固定します。
+
+## 背景
+
+現状の動的な最大値だと、例えば「100ページの本で10ページ読んだ」ときにグラフが満タン（最大値が10）に見えてしまい、全体の進捗を誤認しやすくなります。最初から「100」を最大値として固定しておくことで、全体に対して今どれくらい読み進められているかを直感的に把握できるようにします。
+
+## 実装対象の機能
+
+### 1. グラフの最大値（Y軸の最大値）の固定
+- グラフを表示する際、最大値を示す `@progress_chart_max_pages` を常に `@book.pages` で固定する。
+
+### 2. グラフデータ生成処理の共通化とバグ修正
+- `BooksController` と `BookMemosController` で重複している `prepare_progress_chart_data` 関連処理を共通 Concern `ProgressChartPreparable` に切り出す。
+- `BookMemosController` 側でバリデーションエラー時に `cumulative_pages` が生成されず画面描画でエラーが発生しうるバグを解消する。
+
+## 受け入れ条件
+
+### グラフの最大値の固定
+- [ ] 詳細画面を開いた時点で、進捗グラフの最大値（Y軸の最大値）がその本の「総ページ数」に設定されていること。
+- [ ] 進捗の更新（ページの追加など）を行っても、最大値は総ページ数のままであり、進捗の推移が全体に対する割合として視覚的に分かりやすいこと。
+
+### 共通化とバグ修正
+- [ ] 共通 Concern `ProgressChartPreparable` が正しく作成され、両コントローラから呼び出されていること。
+- [ ] 読書ログがある状態でメモのバリデーションエラーが発生した場合でも、エラー（500）にならず、グラフが正しく描画された状態でエラーメッセージが表示されること。
+
+## 参照ドキュメント
+
+- `docs/product-requirements.md` - プロダクト要求定義書
+- `docs/functional-design.md` - 機能設計書
+- `docs/architecture.md` - アーキテクチャ設計書
