+++ app/controllers/book_memos_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class BookMemosController < ApplicationController
+  include ProgressChartPreparable
+
   before_action :authenticate_user!
   before_action :set_book
   before_action :set_book_memo, only: [ :destroy, :edit, :update ]
@@ -49,37 +51,6 @@ class BookMemosController < ApplicationController
     params.require(:book_memo).permit(:content, :page_number)
   end
 
-  def prepare_progress_chart_data
-    unless @book.reading_logs.exists?
-      @progress_chart_data = []
-      @progress_chart_max_pages = 0
-      return
-    end
-
-    end_date = progress_chart_end_date
-    start_date = progress_chart_start_date(end_date)
-
-    logs_by_date = @book.reading_logs
-                        .where(read_at: start_date..end_date)
-                        .group(:read_at)
-                        .sum(:pages_read)
-
-    @progress_chart_data = (start_date..end_date).map do |date|
-      { date: date, pages_read: logs_by_date.fetch(date, 0) }
-    end
-    @progress_chart_max_pages = @progress_chart_data.map { |row| row[:pages_read] }.max.to_i
-  end
-
-  def progress_chart_end_date
-    return @book.completed_at.to_date if @book.completed? && @book.completed_at.present?
-
-    Date.current
-  end
-
-  def progress_chart_start_date(end_date)
-    first_log_date = @book.reading_logs.minimum(:read_at)
-    start_date = first_log_date || @book.created_at.to_date || end_date
-
-    start_date > end_date ? end_date : start_date
-  end
+  # prepare_progress_chart_data, progress_chart_end_date, and progress_chart_start_date
+  # are provided by ProgressChartPreparable concern.
 end
